Equippable items now also have sprite swap data
So we are all ready to add armor. (But ATM I don't know how to add individual sprites to the sprite library without creating a whole new version of the sprite and copy pasting on the skeleton -- this is mostly fine; you can just import copies of the sprite showing the entire armor "set"). We can also add leg and feet slots now -- but I think that would be bad for the game due to limited inventory space).